Hide users page when user management is disabled (#334)
Cartridge supports custom auth backends. It provides a default one
(which stores users in clusterwide config), but in some cases it may be
implemented differently (via ldap or disabled at all).

When user management is disabled WebUI users page is useless
so it's not displayed. Enabling auth is still a valid case, so auth button
is shown on the main cluster page.

Close #330