Fix 2pc backward compatibility with 1.2.0
To update cartridge without downtime it's necessary to perform rolling
upgrade: restart half a cluster with new version, switch leaders, then
restart the other half.

In #331 we've changed naming of two-phase commit internal functions,
so compatibility between old and new version was broken and it was
impossible to switch the leadership during upgrade.

This patch brings the compatibility back and covers rolling upgrade
scenario wit luatest.