Fix flaky tar test (#590)
Function `tar_execute` is used for unpacking files with `tar -xOf` and
strips the resulting string. One of fixtures ("8.tar") is generated
using `digest.urandom()`, and when it occurs to end with a newline or a
whitespace, `t.assert_equals` fails.

This patch removes implicit string stripping from `tar_execute`.