readme: update intro section for nobbies
Before these changes, README briefly told about the first start of
an application and it was described only at the end of the README.

Users just want to quickly start and try new technology and we need
to explain how to do this. At now, we have a brief section with
"Getting Started" information and also we have a screenshot of the UI.

Screenshot also important for a new user, because he instantly can
understand what we offer.