confapplier: introduce automatic schema upgrade
This patch introduces an option - `upgrade_schema` to perform
an upgrade procedure if we want to upgrade tarantool version.
The main idea to apply box.schema.upgrade on the leader instance.
We can't do it on each instance - it lead to replication
conflicts if we have 2 or more writable instances.

We don't do it automatically to give an ability for user to check
that instances works fine with new tarantool version but old
schema. This thesis is not relevant just now because of
"Space 277 doesn't exist" (tarantool/tarantool#4691).
But I hope it will be solved soon.

Closes #640