Stateful failover (#604)
This patch introduces new failover mode - stateful.

There are 3 main concepts:

**Internal coordinator** is a new role, which makes decisions regarding
leadership. Earlier it was a part of every instances' failover module,
but now it's split away. There may be only one active coordinator in
cluster at a time. Its uniqueness is ensured by external storage which
manages the lock and saves appointments.

**External storage** (`kingdom.lua`) is a stand-alone Tarantool instance
which provides locking mechanism and keeps decisions made by the
coordinator.

**Failover module** (the old one) operates on every instance in cluster
and gathers leadership information for others modules. It was refactored
too, and now it can be described with 4 functions with clearly separated
responsibilities:

- `_get_appointments_*` generates leadership map by itself or polls it
  from external storage depending on the mode setting
  (disabled/eventual/stateful).
- `accept_appointments()` just refreshes the cache and tracks if
  anything changed.
- `failover_loop` (a fiber) repeatedly gets new appointments and accepts
  them using corresponding functions from the above.
- `cfg` is called from `confapplier.apply_config()` (on restart or on
  committing new clusterwide configuration). At first it gets
  appointments synchronously and then starts the failover loop.