Port fetch-schema util to OS X (#688)
"mktemp" on Linux and OS X has different implementations. So, some
features of Linux "mktemp" is unsupported on OS X/FreeBSD. However
such function was used only as random name generator. It was
replaced with "uuidgen" that should be supported on all platforms.

Also this patch runs "graphql-cli" with npx. Sometimes we could
get "graphql: command is not found".

Closes #683