Support automatic schema upgrade (#671)
This patch introduces an option - `upgrade_schema` to perform an upgrade
procedure if we want to upgrade tarantool version. The main idea to
apply `box.schema.upgrade` on the leader (according to failover priority
in topology confgig). We can't do it on each instance - it lead to
replication conflicts if we have 2 or more writable instances.

We don't do it automatically to give an ability for user to check that
instances works fine with new tarantool version but old schema. This
thesis is not relevant just now because of "Space 277 doesn't exist"
error (tarantool/tarantool#4691). But I hope it will be solved soon.

Also introduce new type of tests - "upgrade". The main idea it to try to
bootstrap server from existing snapshots with older schema version.

See also #587 

Co-authored-by: Yaroslav Dynnikov <yaroslav.dynnikov@gmail.com>