cleanup cartridge sources
Non-executable files that contain shebang but doesn't countain
executable bit causes following rpmlint error when we try to pack
cartridge application into rpm:
```
This text file contains a shebang or is located in a path dedicated
for executables, but lacks the executable bits and cannot thus be
executed. If the file is meant to be an executable script, add the
executable bits, otherwise remove the shebang or move the file elsewhere.
```

This patch just removes shebangs from files that located in
cartridge directory.