Fix flaky state_machine test
One of tests checks that instance becomes a leader under some
circumstances, but the check is incorrect. It checks
that instance is writable in several consequent calls, but the instance
may trigger state more quick.

With this patch the check is performed using async call, which tells if
an instance was writable at least for a moment.