Fix flaky state_machine test (#973)
In test_fiber_cancel we check that instance becomes a leader under some
circumstances, but the check isn't correct. It queries if instance is
writable in several consequent calls, but the instance may trigger state
more quickly.

With this patch the check is performed using async call, which tells if
an instance was writable at least for a moment.