Eliminate Gitlab CI pipeline warnings (#1019)
When you use rules with a when: clause without an if: clause, multiple
pipelines may run. Usually this occurs when you push a commit to a
branch that has an open merge request associated with it.

To prevent duplicate pipelines, use workflow: rules or rewrite your
rules to control which pipelines can run.

Since our main repo is github we don't open merge requests on gitlab and
can safely get rid of pipeline warnings.