Override fio.tempdir function in tests (#1059)
Recently in [1] we've monkey-patched `fio.tempdir` implementation
because it didn't respect `$TMPDIR` env variable. In [2] we've extended
it with `base` agrument because etcd required another temp directory
(`/dev/shm` tmpfs in Gitlab CI runner wasn't large enough).  Later
`$TMPDIR` was supported in tarantool (see [3]), but it still doesn't
support passing `base` as an argument, so the cartridge tests fail for
2.4+.

This patch completely eliminates the use of built-in implementation in
favor of ours.

[1] https://github.com/tarantool/cartridge/pull/613
[2] https://github.com/tarantool/cartridge/pull/966
[3] https://github.com/tarantool/tarantool/issues/4794

Part of #1058