Fix remote_control_test for tarantool 2.4+ (#1118)
The Tarantool `box.error` API was significantly refactored in the scope
of [1] and now cartridge tests fail. The problem is that earlier all
errors were converted to ClientError after being passed to the
client-side, and an error code was added to them. Now the error does not
change type after transmission, and if it does not derive from
ClientError, a code field will be missing.

Those changes affect `box` iproto only, but `remote-control` behavior
still corresponds to the old Tarantool versions. 

This patch relieves the checks in `remote-control` test to avoid
unnecessary coupling with tarantool internals.

[1] https://github.com/tarantool/tarantool/issues/4398

Close #1058