Weaken fencing params validation (#1179)
The fencing feature is suitable in stateful mode only, but its config is
always validated. One of the restrictions is
`fencing_timeout < failover_timeout`. But in other failover modes, it
doesn't help and disturbs a user. A user may switch from stateful
failover to eventual and try to change failover timeout. In this case,
any errors regarding fencing would be awkward.