Fix upstream.idle issue tolerance
When there's no workload on a cluster, cartridge usually shows an issue:
"Replication from ... to ...: long idle (1 > 1)".  If the master has no
updates to send to the replicas, it sends heartbeat messages every
`replication_timeout` seconds, and the issue is shown when
`upstream.idle > box.cfg.replication_timeout`.

This patch adds the 2x threshold to avoid unnecessary warnings.