Add implies_storage, implies_router attributes to cluster.known_roles GraphQL
Closes #1071

TDG explicitly marks vshard-storage role as hidden in favor of simple "storage".
WebUI is unable to determine when vshard could be bootstrapped.

WebUI renders "Edit replicaset" dialog using the following query:
{
  cluster {
    known_roles {
      name dependencies
    }
  }
}
Let's extend it with implies_storage, implies_router attributes (both boolean).
It may seem that implies_router is a new concept, but it's not.
It's already used on the frontend implicitly.
It's tied on a hardcoded name of the "vshard-router" role.
Using this option makes it possible to eliminate implicit relations.