Add implies_storage, implies_router attributes to cluster.known_roles GraphQL (#1217)
TDG explicitly marks the `vshard-storage` role as hidden in favor of
simple `storage`. Thus WebUI is unable to determine when vshard could
be bootstrapped.

WebUI renders the "Edit replicaset" dialog using the following query:

```
{
  cluster {
    known_roles { name dependencies }
  }
}
```

This patch extends it with two more attributes - `implies_storage`,
`implies_router` (both boolean).

It may seem that `implies_router` is a new concept, but it's not. It's
already used on the frontend implicitly. It's tied on a hardcoded name
of the "vshard-router" role. Using this option makes it possible to
eliminate implicit relations.

Part of #1071