Take care of errno constants
While working on another patch we encountered weird test behavior.
Occasionally it used to fail with impossible message:

```
t.assert_equals(errno.strerror(errno.ENOPERM), 'Permission denied')

-- ./test/unit/upload_test.lua:178:
-- expected: "Permission denied"
--   actual: "No such file or directory"
```

That was a misspelling - it's actually `EPERM`, but not `ENOPERM`.
But the debugging took a while.

This patch makes work with errno stricter:

- In test helpers, errno list is supplied with asserting metatable.
- In the code, all constants are asserted manually.