Fix replication info in the WebUI (#1284)
There is a bug when a replica with medium id is removed from
`box.space._cluster`. E.g. there are 2 instances and we expel the
leader (switching it beforehand). It results in a gapped array
`box.info.replication` and the GraphQL schema validation fails.

This patch makes the ReplicaStatus nullable in the GraphQL schema
and fixes the response generation accordingly.