Sandbox unit.cfg_errors test (#1285)
Recently we've encountered a weird test behavior: a test (`unit.upload`
from #1266) succeded when run alone, but failed in conjunction with
`unit.cfg_errors` test. The reason is in monkeypatching and reloading of
lua modules used there.

With this patch, all test functions are moved to a simple empty server
and run over netbox (remote-control).