Proxy API calls from and an unconfigured instance (#1278)
Nowadays, every unconfigured instance only displays itself and the other
unconfigured instances. Even if there's a cluster somewhere near, it
isn't shown.

With this patch, a user is able to observe the cluster even from an
unconfigured server. The only requirement is that it's connected to the
membership.

There're still some restrictions regarding managing the cluster. One
can't edit failover params, auth and users, code and scheme, upload /
download config. But the current unconfigured instance can be joined to
the existing cluster. And it doesn't result in another separate cluster
creation anymore.