Avoid WebUI and pool.map_call requests hanging (#1291)
Network connection problems occasionally may result in unexpectedly
long requests processing. Despite all mapped calls are supplied with a
corresponding timeout, it may take much longer. The reason is in the
`connect_timeout` which is 10 seconds by default.

In this patch the use of `pool.connect()` across the code is revised.