Add eventloop speedometer with refreshing membership payload
In case when tarantool node use 100 cpu in tx thread
it's hard to understand is it stuck on some infinite loop or
it's just heavy operation with rarely yields

This eventloop speedometer measure event loop speed. If it is slower
than it have to be, update membership payload.

This way we can understand in membership network is member suspect
because of rarely yields on it.