Fix stateboard longpolling inconsistency (#1384)
The stateboard long-polling algorithm has the bug. It persists
`longpoll_index` in the `box.session.storage` upon the response,
but it doesn't mean the client is able to receive the result. In case
of networking problems, the `longpoll` may result in a temporary error

```
main/131/cartridge.stateful-failover failover.lua:554 W> "localhost:4401": Timeout exceeded
```

but the connection remains alive. The next `longpoll` request
proceeds with the `longpoll_index` already incremented in the
`box.session.storage` and the client never receives the promotion.

With this patch, the `longpoll_index` is passed by the client explicitly.

Both forward and backward compatibility is preserved. It was tested
manually in two ways:

- new stateboard + old cartridge instances;
- old stateboard + new cartridge instances;

In both cases all related tests remain green:

- `integration.switchover.stateboard`
- `integration.failover_stateful.stateboard`
- `integration.fencing.stateboard`