Fix remote-control greeting
The greeting in iproto protocol was used mostly for the feature telling
until now. For the remote-control server it's not correct to restond
with a version from _TARANTOOL variable, because the remote-control
implementation is constant.

This patch locks we the greeting on 1.10.0 as it corresponds to that
feature set.

This patch is necessary so that further tarantool versions could
preserve net.box compatibility with remote-control server.

See also https://github.com/tarantool/tarantool/issues/6253