Fix rebalancer broken by roles reload (#1565)
Stopping the vshard-storage role still isn't implemented
(see https://github.com/tarantool/vshard/issues/121).
So we use nasty workarounds to simulate it.

One of the internal netbox connections wasn't closed. It was reused by
the rebalancer, but roles reload had corrupted it by killing a control
fiber, so the rebalancer got stuck. 

This patch enhances the cleanup and fixes the rebalancer.