Allow to extract filename from http request body (#1613)
This patch introduces the third argument for metadata of http
request parsing (the second is reserved for an error). Before this
patch function http_read_body returned only file content. However
some of our customers want to know filename as well. It's not so
hard to implement because body headers have such info and the main
goal of this patch is to implement needed logic for header
extraction.