rpc: wait until config commit finishes if role is unavailable
There is a raise condition. We could commit config locally but it
could be in progress on some instance. Before this patch user got
unexpected "Role X unavailable" from instance where such role was
assumed.
Solution is an optimistic approach - detect config apply and try
to wait until it will be finished.

Closes #1575