rpc: support is_async for netbox calls
Previously we didn't support it because errors module
didn't it. But currently we are able to do it. After this
patch user will be able to call rpc_call with is_async=true
flag that will be passed directly to netbox.

Follow-up 4f1fbe77b692feacc42618bb1dba121b6f85cfcf