cfg: don't require systemd to use syslog by default
Historically log defaulted to syslog, only if the instance was started
in systemd environment. Sometime after additional check was added because
of #1361. New check, checks if syslog UNIX socket is available. No need
to check for systemd if we already check that syslog socket exists and
even functions.

This commit removes the systemd check from log initialization because:
1) It is kinda hacky (and doesn't work on Centos 7.9 and RHEL 7.9)
2) It is not needed