ci: fix nightly build after etcd action update (#2208)
The regression is introduced in commit 01e0d203b920 ("Use common etcd
action instead of local (#2196)"). The fix follows a way how another
workflow is fixed in commit c4566cdbc527 ("ci: fix etcd tests in
reusable-backend-test.yml (#2202)").

The problem is that the new setup-etcd action sets the `ETCD_PATH`
environment variable on its own. The action sets it as a path to the
directory with the `etcd` executable. However, cartridge testing expects
a path to the executable itself in this variable.

The workflow-level environment variable definition is rewritten by the
action. The fix is to move the environment variable definition down to
the job step level to overwrite action's `ETCD_PATH`.

Reported-by: Yaroslav Dynnikov <yaroslav.dynnikov@gmail.com>